pyproject.toml
[build-system]
requires = ["setuptools>=61.2", "setuptools_scm[toml]>=3.4.1"]
build-backend = "setuptools.build_meta"

[project]
name = "pip-run"
authors = [
	{ name = "Jason R. Coombs", email = "jaraco@jaraco.com" },
]
description = "install packages and run Python with them"
readme = "README.rst"
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: MIT License",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3 :: Only",
]
requires-python = ">=3.9"
dependencies = [
	"autocommand",
	"path >= 15.1",
	"packaging",
	"more_itertools >= 8.3",
	"jaraco.context",
	"platformdirs",
	'tomli; python_version < "3.11"',
	"jaraco.functools >= 3.7",
	"jaraco.env",
	"coherent.deps",
]
dynamic = ["version"]

[project.urls]
Source = "https://github.com/jaraco/pip-run"

[project.optional-dependencies]
test = [
	# upstream
	"pytest >= 6, != 8.1.*",

	# local
	"nbformat",
	"pygments",
	"jaraco.path",
	"jaraco.test >= 5.3",
	"flit-core",
]

doc = [
	# upstream
	"sphinx >= 3.5",
	"jaraco.packaging >= 9.3",
	"rst.linker >= 1.9",
	"furo",
	"sphinx-lint",

	# tidelift
	"jaraco.tidelift >= 1.4",

	# local
]

check = [
	"pytest-checkdocs >= 2.4",
	"pytest-ruff >= 0.2.1; sys_platform != 'cygwin'",
]

cover = [
	"pytest-cov",
]

enabler = [
	"pytest-enabler >= 2.2",
]

type = [
	# upstream
	"pytest-mypy",

	# local
	"cowsay",
]


[project.scripts]
pip-run = "pip_run:run"

[tool.setuptools.packages.find]
exclude = [
	# duplicate exclusions for pypa/setuptools#2688
	"docs",
	"docs.*",
	"tests",
	"tests.*",
	"examples",
]
namespaces = true


[tool.setuptools_scm]
